+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ~~~~~~~~~
 
+0.30.0 (September 2023)
+-----------------------
+
+* Allow setting of rng seed in qulacs backend.
+* Updated pytket version requirement to 1.19.
+
 0.29.0 (June 2023)
 ------------------